+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Changed
+
+- Abort if no cop was picked on `pick` command.
+
 ## 0.1.0
 
 ### Added